api/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-Cors==4.0.1

marshmallow==3.21.1

python-dotenv==1.0.1

pymysql==1.1.1Flask==3.1.3
Flask-SQLAlchemy==3.1.1
flask-cors==5.0.1
python-dotenv==1.0.1
flask-jwt-extended==4.6.0
Flask-Migrate==4.0.7
PyMySQL==1.1.1
